requirements.txt
pandas==1.5.3
Pillow==9.4.0
requests==2.28.2
selenium==4.8.0
tqdm==4.64.1
